refactor: irrelevant methods to utility folder
